src/machine_learning/model.py
# src/machine_learning/model.py
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler

class MLModel:
    def __init__(self):
        self.model = RandomForestClassifier(n_estimators=100)
        self.scaler = StandardScaler()

    def train(self, X_train, y_train):
        X_train_scaled = self.scaler.fit_transform(X_train)
        self.model.fit(X_train_scaled, y_train)

    def predict(self, X):
        X_scaled = self.scaler.transform(X)
        return self.model.predict(X_scaled)
