requirements.txt
PyQt5
matplotlib
numpy
scipy
scikit-learn
pyserial
